requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
PyPDF2>=3.0.1
pydantic[email]>=2.0.0
